rename package to not conflict with existing bitrise mores
