+++ resultprocessing/ftl_result_processor_test.go
@@ -1,4 +1,4 @@
-package step
+package resultprocessing
 
 import (
 	"testing"
